# proof/live-listing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PANTHERA Marketplace Service Hub",
    "version": "2.0.0",
    "description": "x402-paid intelligence APIs for autonomous agents. Call without payment for HTTP 402 payment discovery, then retry with Payment-Signature."
  },
  "servers": [
    {
      "url": "https://hermes-x402-panthera.loca.lt",
      "description": "Stable localtunnel endpoint"
    },
    {
      "url": "https://publisher-jason-aaa-sent.trycloudflare.com",
      "description": "Cloudflare fallback endpoint"
    }
  ],
  "x-payment": {
    "network": "base",
    "chainId": "eip155:8453",
    "asset": "USDC",
    "assetAddress": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
    "recipient": "0x5f03897c6c77dD00F65222A2420a3Cff5507079D",
    "header": "Payment-Signature"
  },
  "paths": {
    "/api/run": {
      "get": {
        "summary": "Google Maps Lead Generator — search businesses by category + location",
        "description": "Google Maps Lead Generator — search businesses by category + location Price: 0.005 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/details": {
      "get": {
        "summary": "Google Maps Place Details — detailed business info by Place ID",
        "description": "Google Maps Place Details — detailed business info by Place ID Price: 0.005 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/serp": {
      "get": {
        "summary": "Mobile SERP Tracker — Google search results with organic, ads, PAA, AI overview",
        "description": "Mobile SERP Tracker — Google search results with organic, ads, PAA, AI overview Price: 0.003 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "Get job listings (Indeed/LinkedIn) with salary + date + proxy metadata",
        "description": "Get job listings (Indeed/LinkedIn) with salary + date + proxy metadata Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reviews/search": {
      "get": {
        "summary": "Search businesses by query + location",
        "description": "Search businesses by query + location Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reviews/{place_id}": {
      "get": {
        "summary": "Fetch Google reviews by Place ID",
        "description": "Fetch Google reviews by Place ID Price: 0.02 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/business/{place_id}": {
      "get": {
        "summary": "Get business details + review summary",
        "description": "Get business details + review summary Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reviews/summary/{place_id}": {
      "get": {
        "summary": "Get review summary stats",
        "description": "Get review summary stats Price: 0.005 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/linkedin/person": {
      "get": {
        "summary": "LinkedIn person profile enrichment",
        "description": "LinkedIn person profile enrichment Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/linkedin/company": {
      "get": {
        "summary": "LinkedIn company profile enrichment",
        "description": "LinkedIn company profile enrichment Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/linkedin/search/people": {
      "get": {
        "summary": "Search LinkedIn people by keywords",
        "description": "Search LinkedIn people by keywords Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/linkedin/company/{id}/employees": {
      "get": {
        "summary": "Find company employees by title",
        "description": "Find company employees by title Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reddit/search": {
      "get": {
        "summary": "Search Reddit posts by keyword",
        "description": "Search Reddit posts by keyword Price: 0.005 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reddit/trending": {
      "get": {
        "summary": "Get trending Reddit posts",
        "description": "Get trending Reddit posts Price: 0.005 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reddit/subreddit/{name}": {
      "get": {
        "summary": "Browse subreddit posts",
        "description": "Browse subreddit posts Price: 0.005 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/reddit/thread/{path}": {
      "get": {
        "summary": "Fetch post comments",
        "description": "Fetch post comments Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/instagram/profile/{username}": {
      "get": {
        "summary": "Instagram profile data",
        "description": "Instagram profile data Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/instagram/posts/{username}": {
      "get": {
        "summary": "Recent Instagram posts",
        "description": "Recent Instagram posts Price: 0.02 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/instagram/analyze/{username}": {
      "get": {
        "summary": "Full Instagram analysis with AI vision",
        "description": "Full Instagram analysis with AI vision Price: 0.15 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/instagram/analyze/{username}/images": {
      "get": {
        "summary": "AI vision analysis of Instagram images",
        "description": "AI vision analysis of Instagram images Price: 0.08 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/instagram/audit/{username}": {
      "get": {
        "summary": "Instagram authenticity audit",
        "description": "Instagram authenticity audit Price: 0.05 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/airbnb/search": {
      "get": {
        "summary": "Search Airbnb listings by location",
        "description": "Search Airbnb listings by location Price: 0.02 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/airbnb/listing/{id}": {
      "get": {
        "summary": "Get detailed Airbnb listing",
        "description": "Get detailed Airbnb listing Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/airbnb/reviews/{listing_id}": {
      "get": {
        "summary": "Get Airbnb listing reviews",
        "description": "Get Airbnb listing reviews Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/airbnb/market-stats": {
      "get": {
        "summary": "Airbnb market statistics",
        "description": "Airbnb market statistics Price: 0.05 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/ecommerce/check": {
      "get": {
        "summary": "E-Commerce Price & Stock Monitor — price, availability, seller, ratings, reviews, SKU/ASIN/GTIN",
        "description": "E-Commerce Price & Stock Monitor — price, availability, seller, ratings, reviews, SKU/ASIN/GTIN Price: 0.002 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/ecommerce/batch": {
      "post": {
        "summary": "Batch check up to 10 product URLs for price/stock intelligence",
        "description": "Batch check up to 10 product URLs for price/stock intelligence Price: 0.002 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/prediction/signal": {
      "get": {
        "summary": "Prediction Market Signal Aggregator — Polymarket odds + Reddit sentiment + divergence scoring",
        "description": "Prediction Market Signal Aggregator — Polymarket odds + Reddit sentiment + divergence scoring Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/prediction/trending": {
      "get": {
        "summary": "Trending prediction-market topics ranked by volume/liquidity",
        "description": "Trending prediction-market topics ranked by volume/liquidity Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/bounty-scan": {
      "get": {
        "summary": "Hermes Bounty Scanner — rank GitHub bounties by payout, competition, safety, and next action",
        "description": "Hermes Bounty Scanner — rank GitHub bounties by payout, competition, safety, and next action Price: 0.05 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/research": {
      "get": {
        "summary": "Multi-source research aggregation",
        "description": "Multi-source research aggregation Price: 0.05 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/trending": {
      "get": {
        "summary": "Trending topics intelligence",
        "description": "Trending topics intelligence Price: 0.01 USDC. Call without Payment-Signature for HTTP 402 discovery and schema.",
        "responses": {
          "200": {
            "description": "Paid request succeeded"
          },
          "402": {
            "description": "Payment required; response includes networks, price, required headers, and output schema"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    }
  }
}